% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-dotProduct.R
\name{adot}
\alias{adot}
\title{Aitchison Scalar Product for Vectors}
\usage{
adot(x, w)
}
\arguments{
\item{x}{A composition of node values.}

\item{w}{A composition of node weights.}
}
\value{
A scalar product.
}
\description{
This function computes the Aitchison scalar product of
 two compositions, defined as the sum of the product
 of all pairwise log-ratios of \code{x} and of \code{w},
 scaled by the number of components.
}
\examples{
x <- randAcomp(20)[1,]
w <- randAcomp(20)[1,]
adot(x, w)
}
