% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-dotProduct.R
\name{amat}
\alias{amat}
\title{Aitchison Scalar Product for Matrices}
\usage{
amat(X, W)
}
\arguments{
\item{X}{A matrix of sample compositions.}

\item{W}{A matrix of weight compositions.}
}
\value{
A matrix of scalar products.
}
\description{
This function computes the Aitchison scalar product of
 two matrices, The first matrix should describe N sample
 compositions (as rows) with D components (as columns),
 while the second matrix should describe D weight
 compositions (as rows) for M outputs (as columns).
}
\examples{
X <- randAcomp(4, 5)
W <- t(randAcomp(3, 5))
amat(X, W)
}
