% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-amalgam.R
\name{as.slr}
\alias{as.slr}
\title{Turn Components into Log-Ratios}
\usage{
as.slr(amalgams)
}
\arguments{
\item{amalgams}{A \code{matrix} or \code{amalgam} object.}
}
\value{
A \code{matrix}.
}
\description{
The \code{as.slr} function turns a composition
 (or \code{amalgam} object) into a set of log-ratios by taking
 the log of the first column over the second, the log of the
 third column over the fourth, and so on.
}
\details{
When the input is an \code{amalgam} object, the result
 is a set of summed log-ratios (SLRs).
}
