% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-nbits.R
\name{int.from2bit}
\alias{int.from2bit}
\title{Convert Bits to Integers}
\usage{
int.from2bit(codon)
}
\arguments{
\item{codon}{A binary vector.}
}
\description{
This function converts bits into integers.
}
\details{
Use \code{\link{int.fromNbit}} instead.
}
\examples{
int.from2bit(c(0,0,1,1,0,0,1,1))
}
