% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-entropy.R
\name{objective.diffEntropy}
\alias{objective.diffEntropy}
\title{Calculate Fitness of Binary Vector}
\usage{
objective.diffEntropy(codon, ARGS)
}
\arguments{
\item{codon}{A binary vector.}

\item{ARGS}{Handled by \code{\link{prepareArgs}}.}
}
\description{
This objective function seeks to maximize the (average)
 between-group difference in the entropy of the amalgamation.
 This function expects exactly two groups.
}
