% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-objectives.R
\name{objective.keepDist}
\alias{objective.keepDist}
\title{Calculate Fitness of Binary Vector}
\usage{
objective.keepDist(codon, ARGS)
}
\arguments{
\item{codon}{A binary vector.}

\item{ARGS}{Handled by \code{\link{prepareArgs}}.}
}
\description{
This objective function seeks to maximize the correlation
 between the Aitchison distance of the complete composition
 and the Aitchison distance of the amalgamation.
}
