% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-entropy.R
\name{objective.keepEntropy}
\alias{objective.keepEntropy}
\title{Calculate Fitness of Binary Vector}
\usage{
objective.keepEntropy(codon, ARGS)
}
\arguments{
\item{codon}{A binary vector.}

\item{ARGS}{Handled by \code{\link{prepareArgs}}.}
}
\description{
This objective function seeks to preserve the relative entropy
 of the amalgamation for all samples.
}
