% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-entropy.R
\name{objective.keepSKL}
\alias{objective.keepSKL}
\title{Calculate Fitness of Binary Vector}
\usage{
objective.keepSKL(codon, ARGS)
}
\arguments{
\item{codon}{A binary vector.}

\item{ARGS}{Handled by \code{\link{prepareArgs}}.}
}
\description{
This objective function seeks to maximize the correlation
 between the (symmetric) Kullback-Leibler divergence of the
 complete composition and the (symmetric) Kullback-Leibler
 divergence of the amalgamation.
}
