% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-entropy.R
\name{objective.maxDSKL}
\alias{objective.maxDSKL}
\title{Calculate Fitness of Binary Vector}
\usage{
objective.maxDSKL(codon, ARGS)
}
\arguments{
\item{codon}{A binary vector.}

\item{ARGS}{Handled by \code{\link{prepareArgs}}.}
}
\description{
This function is adapted from \code{\link{objective.maxRDA}}.
 It uses the PCoA of the SKL divergences to maximize the percent of the
 total variance in the PCoA that can be explained by the matrix z.
 In formula notation, RDA(PCoA ~ z).
}
