% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-objectives.R
\name{objective.maxRDA}
\alias{objective.maxRDA}
\title{Calculate Fitness of Binary Vector}
\usage{
objective.maxRDA(codon, ARGS)
}
\arguments{
\item{codon}{A binary vector.}

\item{ARGS}{Handled by \code{\link{prepareArgs}}.}
}
\description{
This objective function seeks to maximize the percent of the
 total variance present in the matrix z that can be explained
 by the amalgams.
}
