% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-objectives.R
\name{objective.maxRDA2}
\alias{objective.maxRDA2}
\title{Calculate Fitness of Binary Vector}
\usage{
objective.maxRDA2(codon, ARGS)
}
\arguments{
\item{codon}{A binary vector.}

\item{ARGS}{Handled by \code{\link{prepareArgs}}.}
}
\description{
This objective function seeks to maximize the percent of the
 total variance present in the amalgamation that can be explained
 by the matrix z.
}
