% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-amalgam.R
\name{prepareArgs}
\alias{prepareArgs}
\title{Argument Handler Function}
\usage{
prepareArgs(x, n.amalgams = 3, maxiter = ncol(x) * 10, z = NULL,
  objective = objective.keepDist, weights = weight.Nto1, ...)
}
\arguments{
\item{x}{A matrix. The input data. Rows are samples and columns are components.}

\item{n.amalgams}{An intger. How many components the amalgamation should have.}

\item{maxiter}{An integer. How long the genetic algorithm should run.}

\item{z}{A matrix. The constraining matrix. Optional.}

\item{objective}{A function. The objective function. See above.}

\item{weights}{A function. The weights function. See above.}

\item{...}{Arguments passed to \code{GA::ga} function.}
}
\value{
A list of arguments.
}
\description{
This function handles arguments for \code{\link{amalgam}}.
}
