% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-methods.R
\name{print.amalgam}
\alias{print.amalgam}
\title{Show \code{amalgam} Object}
\usage{
\method{print}{amalgam}(x, ...)
}
\arguments{
\item{x}{An \code{amalgam} object.}

\item{...}{Not used.}
}
\description{
This function shows an \code{amalgam} object.
}
