% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{randAcomp}
\alias{randAcomp}
\title{Create Random Acomp Data}
\usage{
randAcomp(nrows, ncols = nrows, lambda = 100)
}
\arguments{
\item{nrows}{An integer. The number of rows in the matrix.}

\item{ncols}{An integer. The number of columns in the matrix.}

\item{lambda}{Argument passed to \code{rpois}.}
}
\value{
A matrix.
}
\description{
This function makes a random matrix by sampling rows
 from a Poisson distribution, then closing them.
}
