% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-entropy.R
\name{SKL}
\alias{SKL}
\title{Calculate Symmetric Kullback-Leibler divergence}
\usage{
SKL(A)
}
\arguments{
\item{A}{A matrix of compositional data.}
}
\value{
A matrix of distances.
}
\description{
This function computes the (symmetric) Kullback-Leibler divergence.
}
