% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-entropy.R
\name{SKL_slow}
\alias{SKL_slow}
\title{Calculate Symmetric Kullback-Leibler divergence}
\usage{
SKL_slow(A)
}
\arguments{
\item{A}{A matrix of compositional data.}
}
\value{
A matrix of distances.
}
\description{
This function computes the Symmetric Kullback-Leibler divergence.
}
