% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-wadist.R
\name{wadist}
\alias{wadist}
\title{Calculate Weighted Aitchison Distance}
\usage{
wadist(A)
}
\arguments{
\item{A}{A matrix of compositional data.}
}
\value{
A matrix of distances.
}
\description{
This function computes the weighted aitchison distance proposed by
 Greenacre & Lewi (2009) in \code{doi:10.1007/s00357-009-9027-y}.
 Unlike the unweighted Aitchison distance, the weighted Aitchison
 distance has distributional equivalence.
}
