% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-weights.R
\name{weight.Nto1}
\alias{weight.Nto1}
\title{Convert Bits to Weights Matrix}
\usage{
weight.Nto1(codon, n.amalgams)
}
\arguments{
\item{codon}{A binary vector.}

\item{n.amalgams}{The number of columns in the weights matrix.}
}
\description{
This function builds a weights matrix from a binary vector.
 The resultant weights matrix with have \code{n.amalgams} columns
 and a number of rows based on the size of the binary vector.
 The resultant weights matrix will have at most 1 value per row.
 It is possible for row sums or column sums to equal 0.
}
\details{
The interpretation of this weights matrix is that each component
 will only contribute to one amalgam node.
}
\examples{
# For 3 amalgams, read as "node 1", "node 3", "node 2", "node 0"
weight.Nto1(c(0,1,1,1,1,0,0,0), 3)
}
