% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-weights.R
\name{weight.NtoN}
\alias{weight.NtoN}
\title{Convert Bits to Weights Matrix}
\usage{
weight.NtoN(codon, n.amalgams)
}
\arguments{
\item{codon}{A binary vector.}

\item{n.amalgams}{The number of columns in the weights matrix.}
}
\description{
This function builds a weights matrix from a binary vector.
 The resultant weights matrix with have \code{n.amalgams} columns
 and a number of rows based on the size of the binary vector.
 The resultant weights matrix will have rows that sum to 1.
 It is possible for row sums or column sums to equal 0.
}
\details{
The interpretation of this weights matrix is that each component
 can contribute to multiple amalgam nodes, but must contribute
 an equal amount to each node.
}
\examples{
# For 2 amalgams, read as "node 2", "node 1+2", "node 1", "node 0"
weight.NtoN(c(0,1,1,1,1,0,0,0), 2)
}
