% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-methods.R
\name{plot.amalgam}
\alias{plot.amalgam}
\title{Plot \code{amalgam} Object}
\usage{
\method{plot}{amalgam}(x, col = rep(1, nrow(x$amalgams)),
  center = FALSE, scale = FALSE, a1 = 1, a2 = 2, a3 = 3, ...)
}
\arguments{
\item{x}{An \code{amalgam} object.}

\item{col}{A character vector. Used to color the figure.}

\item{center}{A logical. Toggles whether to center the ternary plot.}

\item{scale}{A logical. Toggles whether to scale the ternary plot.}

\item{a1, a2, a3}{The amalgams to show in the ternary plot.}

\item{...}{Not used.}
}
\description{
This function plots an \code{amalgam} object.
}
